% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.R
\name{inc}
\alias{inc}
\title{Generate a section for the yaml input}
\usage{
inc(input, sep = "\\n  ")
}
\arguments{
\item{input}{a file containing markdown text}

\item{sep}{a separator for each line. Defaults to "\\n  "}
}
\value{
a string
}
\description{
Generate a section for the yaml input
}
\examples{
f <- file()
cat("  this is\\nsome text that\\nwill be renedered in\\na file\\n", file = f)
cat(inc(f))
close(f)
}

