% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monashthesisdown.R
\docType{package}
\name{monashthesisdown}
\alias{monashthesisdown}
\title{monashthesisdown: A package for creating undergrad and post-grad theses using RMarkdown}
\description{
This package inspired by thesisdown and  Prof’ Rob J Hyndman’s work, uses bookdown package  knits your thesis in pdf, gitbook and epub format according to the guidelines of The Monash University
}
\details{
It is advisable that you load **bookdown**, along with **monashthesisdown**
}
