% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.R
\name{thesis_epub}
\alias{thesis_epub}
\title{Creates an R Markdown epub Thesis document}
\usage{
thesis_epub(...)
}
\arguments{
\item{...}{additional arguments passed to \code{bookdown::epub_book}}
}
\value{
A ebook version of the thesis
}
\description{
This is a function called in output in the YAML of the driver Rmd file
to specify the creation of a epub version of the thesis.
}
\examples{
\dontrun{
 output: unimelbdown::thesis_epub
}
}
