% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.R
\name{thesis_pdf}
\alias{thesis_pdf}
\title{Creates an R Markdown PDF Thesis document}
\usage{
thesis_pdf(toc = TRUE, toc_depth = 3, highlight = "default", ...)
}
\arguments{
\item{toc}{A Boolean (TRUE or FALSE) specifying whether table of contents
should be created}

\item{toc_depth}{A positive integer}

\item{highlight}{Syntax highlighting style. Supported styles include
"default", "tango", "pygments", "kate", "monochrome", "espresso", "zenburn",
and "haddock". Pass NULL to prevent syntax highlighting.}

\item{...}{additional arguments passed to \code{bookdown::pdf_book}}
}
\value{
A modified \code{pdf_document} based on the Reed Senior Thesis LaTeX
  template
}
\description{
This is a function called in output in the YAML of the driver Rmd file
to specify the LaTeX template and cls files.
}
\examples{
\dontrun{
 output: unimelbdown::thesis_pdf
}

}
