% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.R
\name{thesis_word}
\alias{thesis_word}
\title{Creates an R Markdown Word Thesis document}
\usage{
thesis_word(...)
}
\value{
A Word Document based on (hopefully soon, but not currently)
the Reed Senior Thesis Word template
}
\description{
This is a function called in output in the YAML of the driver Rmd file
to specify the creation of a Microsoft Word version of the thesis.
}
\examples{
\dontrun{
 output: thesisdown::thesis_word
}
}

